[User]-Refactor apiClient
