+++ src/pages/users/components/EditUserModal.tsx
@@ -9,7 +9,7 @@ import { useForm, SubmitHandler } from "react-hook-form";
 import { zodResolver } from '@hookform/resolvers/zod';
 import { z } from 'zod';
 import { UserProfile } from "@/types/userProfile.type";
-import { uploadToCloudinary } from "@/lib/apiClient";
+import { uploadToCloudinary } from "@/lib/apiUser";
 import AvatarSection from "../../../components/modal/components/AvatarSection";
 import UserProfileForm from "../../../components/modal/components/UserProfileForm";
